typings update in progress, integrating expressions and queries
