+++ packages/sqorn-pg/types/buildable.d.ts
@@ -1,0 +1,5 @@
+export interface Buildable {
+  _build(ctx: any): string;
+  query: { text: string, args: any[] };
+  unparameterized: string;
+}
