+++ packages/sqorn-pg/types/expressions/operations/json.d.ts
@@ -1,0 +1,4 @@
+import { JSONTypes } from '..'
+
+export interface JSONOperations {}
+export interface JSONChainOperations<T extends JSONTypes> {}
