+++ packages/sqorn-pg/types/expressions/operations/row.d.ts
@@ -1,0 +1,4 @@
+import { RowTypes } from '..'
+
+export interface RowOperations {}
+export interface RowChainOperations<T extends RowTypes> {}
